refactor: store imports to avoid circular deps
