+++ frontend/src/components/Buttons/Playback/NextTrackButton.vue
@@ -11,8 +11,7 @@
 </template>
 
 <script setup lang="ts">
-import { playbackManagerStore } from '@/store';
+import { playbackManager } from '@/store/playbackManager';
 
 const props = defineProps<{ size?: string }>();
-const playbackManager = playbackManagerStore();
 </script>
