+++ frontend/src/components/Buttons/SubtitleSelectionButton.vue
@@ -37,16 +37,15 @@
 </template>
 
 <script setup lang="ts">
+import { playbackManager } from '@/store/playbackManager';
 import { SubtitleDeliveryMethod } from '@jellyfin/sdk/lib/generated-client';
+import IMdiCheck from 'virtual:icons/mdi/check';
 import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
-import IMdiCheck from 'virtual:icons/mdi/check';
-import { playbackManagerStore } from '@/store';
 
 const menuModel = defineModel<boolean>();
 
 const { t } = useI18n();
-const playbackManager = playbackManagerStore();
 
 const tracks = computed(() => {
   const subs = playbackManager.currentItemParsedSubtitleTracks;
