+++ frontend/src/components/Playback/UpNext.vue
@@ -68,15 +68,13 @@
 </template>
 
 <script setup lang="ts">
-import { playbackManagerStore } from '@/store';
+import { playbackManager } from '@/store/playbackManager';
 import { getEndsAtTime, getRuntimeTime } from '@/utils/time';
 import { computed, ref, watch } from 'vue';
 
 const emit = defineEmits<{
   change: [isVisible: boolean];
 }>();
-
-const playbackManager = playbackManagerStore();
 
 const isHiddenByUser = ref(false);
