+++ frontend/src/components/System/LocaleSwitcher.vue
@@ -28,7 +28,7 @@
 </template>
 
 <script setup lang="ts">
-import { clientSettingsStore } from '@/store';
+import { clientSettings } from '@/store/clientSettings';
 import { getLocaleNativeName } from '@/utils/i18n';
 import { useI18n } from 'vue-i18n';
 
@@ -40,7 +40,6 @@ defineProps<{
 }>();
 
 const i18n = useI18n();
-const clientSettings = clientSettingsStore();
 </script>
 
 <style lang="scss" scoped>
