+++ frontend/src/plugins/router/middlewares/playback.ts
@@ -1,6 +1,6 @@
+import { playbackManager } from '@/store/playbackManager';
 import { isNil } from 'lodash-es';
 import { RouteLocationNormalized, RouteLocationRaw } from 'vue-router/auto';
-import { playbackManagerStore } from '@/store';
 
 /**
  * Validates that no playback is happening when accesing a route
@@ -8,8 +8,6 @@ import { playbackManagerStore } from '@/store';
 export default function playbackGuard(
   to: RouteLocationNormalized
 ): boolean | RouteLocationRaw {
-  const playbackManager = playbackManagerStore();
-
   if (to.path.includes('playback') && isNil(playbackManager.currentItem)) {
     return { path: '/', replace: true };
   }
