+++ frontend/src/utils/store-sync.ts
@@ -1,8 +1,8 @@
+import { usei18n, useRemote, useSnackbar } from '@/composables';
+import { taskManager } from '@/store/taskManager';
 import { DisplayPreferencesDto } from '@jellyfin/sdk/lib/generated-client';
 import { getDisplayPreferencesApi } from '@jellyfin/sdk/lib/utils/api/display-preferences-api';
 import { destr } from 'destr';
-import { usei18n, useRemote, useSnackbar } from '@/composables';
-import { taskManagerStore } from '@/store';
 
 const CLIENT = 'vue';
 
@@ -173,7 +173,6 @@ export async function syncCustomPrefs<T extends object>(
   customPrefs: T
 ): Promise<void> {
   const { t } = usei18n();
-  const taskManager = taskManagerStore();
 
   /**
    * Creates a config syncing task, so UI can show that there's a syncing in progress
